+++ apps/raw_material/views.py
@@ -9,22 +9,39 @@ from apps.raw_material.models import Lot, ILot
 from apps.raw_material.serializers import LotListSerializer, LotSerializer, LotDetailSerializer, ILotSerializer, \
     ILotCRUDSerializer
 from apps.report.models import Report
+from apps.util.pagination import SetPagination
 
 
 # Create your views here.
 class ListCreateLotView(APIView):
+
     def get(self, request, format=None):
-        if Lot.objects.all().exists():
-            serializer = LotListSerializer(Lot.objects.all(), many=True)
-            return Response({'result': serializer.data}, status=status.HTTP_200_OK)
+        queryset = Lot.objects.all()
+        lot = request.query_params.get('lot', None)
+        category = request.query_params.get('category', None)
+        variety = request.query_params.get('variety', None)
+        condition = request.query_params.get('condition', None)
+        if lot:
+            queryset = queryset.filter(lot__icontains=lot)
+        if category:
+            queryset = queryset.filter(category__name__icontains=category)
+        if variety:
+            queryset = queryset.filter(variety__icontains=variety)
+        if condition:
+            queryset = queryset.filter(condition__icontains=condition)
+        if queryset.exists():
+            paginator = SetPagination()
+            results = paginator.paginate_queryset(queryset, request)
+            serializer = LotListSerializer(results, many=True)
+            return paginator.get_paginated_response({'result': serializer.data})
         else:
             return Response({'error': 'No se encontraron registro de lotes'},
                             status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
     def post(self, request, format=None):
-        # if request.user.role != "6":
-        #     return Response({'error': 'No tiene permisos para realizar esta acción'},
-        #                     status=status.HTTP_401_UNAUTHORIZED)
+        if request.user.role != "6":
+            return Response({'error': 'No tiene permisos para realizar esta acción'},
+                            status=status.HTTP_401_UNAUTHORIZED)
         try:
             serializer = LotSerializer(data=request.data)
             serializer.is_valid(raise_exception=True)
@@ -57,9 +74,9 @@ class DetailEntryView(APIView):
 
     def delete(self, request, *args, **kwargs):
         entry = get_object_or_404(Lot, lot=kwargs['lot'])
-        # if not request.user.is_superuser:
-        #     return Response({'error': 'No tiene permisos para realizar esta acción'},
-        #                     status=status.HTTP_401_UNAUTHORIZED)
+        if not request.user.is_superuser:
+            return Response({'error': 'No tiene permisos para realizar esta acción'},
+                            status=status.HTTP_401_UNAUTHORIZED)
         if entry.closed:
             return Response({"error": "El lote ya esta bloqueado para su edición. Contáctese con el administrador"},
                             status=status.HTTP_500_INTERNAL_SERVER_ERROR)
@@ -84,9 +101,9 @@ class ListILotView(APIView):
 
 class CreateInformationView(APIView):
     def post(self, request, format=None):
-        # if request.user.role != "6":
-        #     return Response({'error': 'No tiene permisos para realizar esta acción'},
-        #                     status=status.HTTP_401_UNAUTHORIZED)
+        if request.user.role != "6":
+            return Response({'error': 'No tiene permisos para realizar esta acción'},
+                            status=status.HTTP_401_UNAUTHORIZED)
         data = request.data
         try:
             if Lot.objects.get(id=data.get("lot")).closed:
@@ -111,9 +128,9 @@ class DetailInformationView(APIView):
         if inf.lot.closed:
             return Response({"error": "El lote ya esta bloqueado para su edición. Contáctese con el administrador"},
                             status=status.HTTP_500_INTERNAL_SERVER_ERROR)
-        # if request.user.role != "6":
-        #     return Response({'error': 'No tiene permisos para realizar esta acción'},
-        #                     status=status.HTTP_401_UNAUTHORIZED)
+        if request.user.role != "6":
+            return Response({'error': 'No tiene permisos para realizar esta acción'},
+                            status=status.HTTP_401_UNAUTHORIZED)
         try:
             if not request.data.get("indicted"):
                 request.data["indicted"] = False
@@ -133,9 +150,9 @@ class DetailInformationView(APIView):
         if inf.lot.closed:
             return Response({"error": "El lote ya esta bloqueado para su edición. Contáctese con el administrador"},
                             status=status.HTTP_500_INTERNAL_SERVER_ERROR)
-        # if not request.user.is_superuser:
-        #     return Response({'error': 'No tiene permisos para realizar esta acción'},
-        #                     status=status.HTTP_401_UNAUTHORIZED)
+        if not request.user.is_superuser:
+            return Response({'error': 'No tiene permisos para realizar esta acción'},
+                            status=status.HTTP_401_UNAUTHORIZED)
         try:
             inf.delete()
             return Response({"message": "Item eliminado correctamente"}, status=status.HTTP_200_OK)
